requirements.txt
flask==3.0.0
gunicorn==21.2.0
openai==1.12.0
httpx==0.27.0
sendgrid==6.11.0
python-dotenv==1.0.0
